feat(rfe375): implement Phase 2 - API versioning and module split
Phase 2 of RFE375 codebase enhancements:

P2.1: API Versioning
- Change proto package from 'flame' to 'flame.v1'
- Update all Rust imports to use flame::v1 namespace
- Update go_package to github.com/flame-sh/flame/sdk/go/rpc/v1

P2.2: Split apis.rs (1350 lines) into focused modules
- common/src/apis/types.rs (411 lines) - domain type definitions
- common/src/apis/session.rs (154 lines) - Session impl blocks
- common/src/apis/to_rpc.rs (374 lines) - domain to RPC conversions
- common/src/apis/from_rpc.rs (400 lines) - RPC to domain conversions
- common/src/apis/mod.rs (63 lines) - re-exports and tests

P2.3: Add code coverage to CI
- Add cargo-tarpaulin installation step
- Add coverage report generation
- Add Codecov upload action

Review feedback addressed:
- Add empty string check in parse_memory() to prevent panic
- Use unwrap_or_default() for timestamp conversion
- Optimize pop_pending_task() to avoid cloning HashMap

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
