+++ .github/workflows/code-verify.yaml
@@ -32,3 +32,13 @@ jobs:
         run: cargo build --release
       - name: Cargo test
         run: cargo test --workspace --exclude flame-rs --exclude cri-rs
+      - name: Install cargo-tarpaulin
+        run: cargo install cargo-tarpaulin
+      - name: Generate coverage report
+        run: cargo tarpaulin --workspace --exclude flame-rs --exclude cri-rs --out xml --output-dir coverage
+      - name: Upload coverage to Codecov
+        uses: codecov/codecov-action@v4
+        with:
+          files: coverage/cobertura.xml
+          fail_ci_if_error: false
+          token: ${{ secrets.CODECOV_TOKEN }}
