+++ rpc/protos/types.proto
@@ -1,8 +1,8 @@
 syntax = "proto3";
 
-package flame;
+package flame.v1;
 
-option go_package = "github.com/flame-sh/flame/sdk/go/rpc";
+option go_package = "github.com/flame-sh/flame/sdk/go/rpc/v1";
 
 message Metadata {
   string id = 1;
