+++ rpc/src/lib.rs
@@ -12,5 +12,7 @@ limitations under the License.
 */
 
 pub mod flame {
-    tonic::include_proto!("flame");
+    pub mod v1 {
+        tonic::include_proto!("flame.v1");
+    }
 }
