+++ sdk/python/protos/frontend.proto
@@ -2,9 +2,9 @@ syntax = "proto3";
 
 import "types.proto";
 
-package flame;
+package flame.v1;
 
-option go_package = "github.com/flame-sh/flame/sdk/go/rpc";
+option go_package = "github.com/flame-sh/flame/sdk/go/rpc/v1";
 
 /*
  * The frontend service of Flame, which is used to communicate with client sdk,
@@ -19,6 +19,10 @@ service Frontend {
   rpc ListApplication(ListApplicationRequest) returns (ApplicationList) {}
 
   rpc ListExecutor(ListExecutorRequest) returns (ExecutorList) {}
+
+  // Node operations
+  rpc ListNodes(ListNodesRequest) returns (NodeList) {}
+  rpc GetNode(GetNodeRequest) returns (GetNodeResponse) {}
 
   rpc CreateSession (CreateSessionRequest) returns (Session) {}
   rpc DeleteSession (DeleteSessionRequest) returns (Session) {}
@@ -61,6 +65,21 @@ message ListApplicationRequest {
 
 message ListExecutorRequest {
   
+}
+
+// ListNodesRequest is the request for listing all registered nodes.
+message ListNodesRequest {
+  // No pagination for now.
+}
+
+// GetNodeRequest is the request for getting a specific node by name.
+message GetNodeRequest {
+  string name = 1;
+}
+
+// GetNodeResponse is the response containing the requested node.
+message GetNodeResponse {
+  Node node = 1;
 }
 
 message CreateSessionRequest {
